\name{annotate}
\alias{annotate}
\title{Applies annotations to variants for informed clustering}
\usage{
  annotate(rv, annotations = NA, file = NA)
}
\arguments{
  \item{rv}{rvclustobject}
}
\value{
  annotated rvclustobject
}
\description{
  ----------------------------------------------------------------------
  '#\cr Annotations allow RVCLUST to estimate which
  features are likely to have \cr similar effects without
  looking at the outcomes in advance. This \cr preserves
  much of the statistical integrity lost with supervised
  \cr learning. These annotations may also aid in
  interpreting results. \cr
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

