\name{burden.test}
\alias{burden.test}
\title{Collapses features into composite feature using the OR operator}
\usage{
  burden.test(rv)
}
\arguments{
  \item{rv}{rvclustobject}
}
\value{
  rvclustobject with collapsed data (composite features)
}
\description{
  ----------------------------------------------------------------------
  '#\cr Performs a standard burden test, which collapses
  the features in a bin \cr by taking the SUM of all
  features. For each subject, the composite \cr feature is
  represented by the burden of events occurring in the bin.
  \cr
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

