\name{collapsing.test}
\alias{collapsing.test}
\title{Collapses features into composite feature using the OR operator}
\usage{
  collapsing.test(rv)
}
\arguments{
  \item{rv}{rvclustobject}
}
\value{
  rvclustobject with collapsed data
}
\description{
  ----------------------------------------------------------------------
  '#\cr Performs a standard collapsing test, which
  collapses the features in a \cr bin by taking the OR of
  all features. For each subject, the composite \cr feature
  is represented as positive if any of the individual
  events \cr have occurred. \cr
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

