\name{single.marker.lm}
\alias{single.marker.lm}
\title{Performs single-marker linear regression on variants}
\usage{
  single.marker.lm(rv, label.by = NA, phen = "PHENOTYPE")
}
\arguments{
  \item{rv}{rvclustobject}
}
\value{
  rvclustobject with variant-level linear regression
  results
}
\description{
  See above
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

