\name{test.aggregation}
\alias{test.aggregation}
\title{Automated test for aggregation modules}
\usage{
  test.aggregation(AGGREGATION.FUN)
}
\arguments{
  \item{aggregation}{function}
}
\value{
  validity
}
\description{
  ----------------------------------------------------------------------
  '#\cr Use this function to test aggregation modules.
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

