\name{test.annotate}
\alias{test.annotate}
\title{Automated test for annotation modules}
\usage{
  test.annotate(ANNOTATE.FUN)
}
\arguments{
  \item{annotation}{function}
}
\value{
  validity
}
\description{
  ----------------------------------------------------------------------
  '#\cr Use this function to test annotation modules.
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

