\name{test.association}
\alias{test.association}
\title{Automated test for statistical association modules}
\usage{
  test.association(ASSOCIATION.FUN)
}
\arguments{
  \item{association}{function}
}
\value{
  validity
}
\description{
  ----------------------------------------------------------------------
  '#\cr Use this function to test statistical association
  modules.
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

