\name{test.cluster}
\alias{test.cluster}
\title{Automated test for clustering modules}
\usage{
  test.cluster(CLUSTER.FUN)
}
\arguments{
  \item{clustering}{function}
}
\value{
  validity
}
\description{
  ----------------------------------------------------------------------
  '#\cr Use this function to test cluster modules.
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}

