\name{annotate}
\alias{annotate}
\title{Uses GWAR to annotate genetic variants}
\usage{
  annotate(rv.dat, annotations = NA)
}
\arguments{
  \item{rv.dat}{rare variant data frame}
}
\value{
  annotated rare variant data frame
}
\description{
  Uses GWAR to annotate genetic variants
}
\examples{
init(rv.dat)
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}
\seealso{
  \code{\link{init}}
}

