\docType{package}
\name{rvclust-package}
\alias{rvclust}
\alias{rvclust-package}
\title{RVCLUST: Clustering rare variants for statistical analysis}
\description{
  Clustering rare variants for statistical analysis
}
\details{
  \tabular{ll}{ Package: \tab rvclust\cr Type: \tab
  Package\cr Version: \tab 2.0\cr Date: \tab 2013-06-01\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }

  ----------------------------------------------------------------------
  '#\cr RVCLUST is a modular framework for aggregating rare
  event features. \cr These features encode events whose
  low frequency limits our ability \cr to detect their
  effects with statistical significance. \cr \cr
  Annotation-based clustering attempts to identify groups
  of features \cr likely to exhibit a similar effect on an
  outcome. These groups are \cr the collapsed to generate
  aggregate features with a higher frequency \cr of
  variation than their constituent members. \cr
  ----------------------------------------------------------------------
  '#\cr
}
\author{
  R Michael Sivley \email{mike.sivley@vanderbilt.edu}
}
\references{
  Sivley, R. Michael, Fish, Alexandra E., Bush, William S.
  (2013).  Knowledge-constrained K-medoids Clustering of
  Regulatory Rare Alleles for Burden Tests. Lecture Notes
  in Computer Science, vol. 7833, pg 35-42.
}
\seealso{
  \code{\link{rvclustobject}}
}
\keyword{package}

